docs: SegmentedBar revamp
