--- app/ng-ui-widgets-category/segmented-bar/basic-segmented-bar/basic-segmented-bar.component.html
@@ -1,6 +1,0 @@
-<StackLayout sdkExampleTitle sdkToggleNavButton>
-	<!-- >> segmentedbar-basic-html -->
-	<SegmentedBar [items]="myItems" selectedIndex="0" (selectedIndexChange)="onSelectedIndexChange($event)" class="m-5"></SegmentedBar>
-	<Label [text]="prop" textWrap="true" verticalAlign="center" class="h2 p-15 text-center"></Label>
-	<!-- << segmentedbar-basic-html -->
-</StackLayout>
