+++ app/ng-ui-widgets-category/segmented-bar/end.md
@@ -1,8 +1,29 @@
+## Properties
 
-**API Reference for the** [SegmentedBar Class](http://docs.nativescript.org/api-reference/modules/_ui_segmented_bar_.html)
+| Name     | Type    | Description    |
+|----------|---------|----------------|
+| `selectedIndex`   | `number` | Gets or sets the selected index of the SegmentedBar component. |
+| `items`   | `Array<SegmentedBarItem>` | Gets or sets the items of the SegmentedBar. |
+| `selectedBackgroundColor`   | `Color` | Gets or sets the selected background color of the SegmentedBar component. |
 
-**Native Component**
+## Events
 
-| Android                | iOS      |
-|:-----------------------|:---------|
+| Name     | Description    |
+|----------|----------------|
+| `selectedIndexChanged`  | Emitted when the selected index is changed (via interaction or programmatically). |
+
+
+## API References
+
+| Name     | Type    | 
+|----------|---------|
+| [tns-core-modules/ui/segmented-bar](http://docs.nativescript.org/api-reference/modules/_ui_segmented_bar_.html) | `Module` | 
+| [SegmentedBar](https://docs.nativescript.org/api-reference/classes/_ui_segmented_bar_.segmentedbar) | `Class` | 
+| [SegmentedBarItem](https://docs.nativescript.org/api-reference/classes/_ui_segmented_bar_.segmentedbaritem) | `Class` |
+| [SelectedIndexChangedEventData](https://docs.nativescript.org/api-reference/interfaces/_ui_segmented_bar_.selectedindexchangedeventdata) | `Interface` |
+
+## Native Component
+
+| Android               | iOS      |
+|:----------------------|:---------|
 | [android.widget.TabHost](http://developer.android.com/reference/android/widget/TabHost.html) | [UISegmentedControl](https://developer.apple.com/library/prerelease/ios/documentation/UIKit/Reference/UISegmentedControl_Class/index.html) |
