+++ app/ng-ui-widgets-category/segmented-bar/overview.md
@@ -1,7 +1,2 @@
-Using a SegmentedBar inside an Angular project gives you a simple way to define a collection of tabbed views.
-The SegmentedBar’s `selectedIndexChange` property notifies you for every changes of the component’s `selectedIndex`.
-
-The specific `SegmentedBar` properties are:
- - `items`
- - `selectedIndex`
- - `selectedIndexChange`
+Using a `SegmentedBar` inside a NativeScript project gives you a simple way to define a collection of tabbed views.
+The SegmentedBar’s `selectedIndexChange` event is emitted you for every changes of the component’s `selectedIndex`.
