--- app/ng-ui-widgets-category/segmented-bar/segmented-bar-views/article.md
@@ -1,5 +1,0 @@
-HTML
-<snippet id='segmentedbar-show-views'/>
-
-Setting SegmentedBar's items and visibility of the views
-<snippet id='segmentedbar-items-setting-visibility'/>
