+++ app/ng-ui-widgets-category/segmented-bar/styling/article.md
@@ -1,0 +1,3 @@
+<snippet id='segmentedbar-show-views'/>
+
+> **Note:** The property `selectedBackgroundColor` will behave differently on different platforms. On Android, it will apply color for the underline of the selected bar item, while on iOS, it will apply background color for the whole selected bar item.
