+++ app/ng-ui-widgets-category/segmented-bar/styling/styling.component.html
@@ -1,0 +1,9 @@
+<StackLayout sdkExampleTitle sdkToggleNavButton>
+	<!-- >> segmentedbar-show-views -->
+	<SegmentedBar [items]="items"
+				  backgroundColor="green" color="white" 
+				  selectedBackgroundColor="red">
+	</SegmentedBar>
+	<!-- << segmentedbar-show-views -->
+</StackLayout>
+§
