+++ app/ng-ui-widgets-category/segmented-bar/usage/article.md
@@ -1,5 +1,2 @@
-HTML
 <snippet id='segmentedbar-basic-html'/>
-
-TypeScript
 <snippet id='add-segmentedbar-items'/>
