+++ app/ng-ui-widgets-category/segmented-bar/usage/usage.component.html
@@ -1,0 +1,8 @@
+<StackLayout sdkExampleTitle sdkToggleNavButton>
+	<!-- >> segmentedbar-basic-html -->
+	<SegmentedBar [items]="mySegmentedBarItems" 
+				  selectedIndex="0" 
+				  (selectedIndexChange)="onSelectedIndexChange($event)">
+	</SegmentedBar>
+	<!-- << segmentedbar-basic-html -->
+</StackLayout>
